moved functions to output helper fle
