+++ src/__tests__/generateConfluenceMarkup.test.ts
@@ -1,5 +1,5 @@
 import * as mariner from '../mariner/index';
-import { cleanMarkup, calculateAgeInWholeDays } from '../Utilities/generateConfluenceMarkup';
+import { removeBracesAndBrackets, calculateAgeInWholeDays } from '../Utilities/outputHelpers';
 import { Issue } from '../issueFinder';
 import { DateTime, Duration } from 'luxon';
 
@@ -109,12 +109,9 @@ describe('generateConfluenceMarkup function', () => {
         const dependency = 'Badges/shields';
         singleIssue[0].createdAt = '2021-01-02T10:22:41Z'; // old issue
 
-        const now = DateTime.utc();
-        const date = now.toISO();
-
         mockDependencyMap.set(dependency, singleIssue);
         const results = mariner.generateConfluenceMarkup(mockDependencyMap);
-        expect(results).toContain(`h2. Updated: ${date}`);
+
         expect(results).not.toContainEqual(`h3. ${dependency}`);
         expect(results).not.toContainEqual('\n||*Title*||*Age*||');
         expect(results).not.toContainEqual(singleIssue[0].title);
@@ -125,27 +122,27 @@ describe('cleanMarkup function', () => {
     it('should return string without altering it', () => {
         const title1 =
             '|[Docs: Improve mocks section for promise-based fs/promises | update docs & jest|';
-        const cleanedMarkdown = cleanMarkup(title1);
+        const cleanedMarkdown = removeBracesAndBrackets(title1);
         expect(cleanedMarkdown).toEqual(
             '|(Docs: Improve mocks section for promise-based fs/promises | update docs & jest|'
         );
     });
     it('should remove a set of curly braces', () => {
         const title = '|{WIP} updating frontend components|';
-        const cleanedMarkdown = cleanMarkup(title);
+        const cleanedMarkdown = removeBracesAndBrackets(title);
         expect(cleanedMarkdown).toEqual('|(WIP) updating frontend components|');
     });
     it('should remove all curly braces', () => {
         const title =
             '|{new babel} teardown do not fail tests in non-watch mode - {imports are removed}|';
-        const cleanedMarkdown = cleanMarkup(title);
+        const cleanedMarkdown = removeBracesAndBrackets(title);
         expect(cleanedMarkdown).toEqual(
             '|(new babel) teardown do not fail tests in non-watch mode - (imports are removed)|'
         );
     });
     it('should remove square brackets', () => {
         const title = '|[[es-lint] resolves deprecated code {updates}|';
-        const cleanedMarkdown = cleanMarkup(title);
+        const cleanedMarkdown = removeBracesAndBrackets(title);
         expect(cleanedMarkdown).toEqual('|((es-lint) resolves deprecated code (updates)|');
     });
 });
